% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cde_gest}
\alias{cde_gest}
\title{Controlled Direct Effect (CDE) of the G-estimation model.}
\usage{
cde_gest(x = NULL, m = NULL, c_xy = NULL, c_my = NULL, y = NULL,
  mX = NULL, mM = NULL, mY = NULL, famX = "gaussian",
  famM = "gaussian", famY = "gaussian")
}
\arguments{
\item{x}{: (datatype) Description of structure. What it represents.}

\item{m}{: (datatype) Description of structure. What it represents.}

\item{c_xy}{: (datatype) Description of structure. What it represents.}

\item{c_my}{: (datatype) Description of structure. What it represents.}

\item{y}{: (datatype) Description of structure. What it represents.}

\item{mX}{: (datatype) Description of structure. What it represents.}

\item{mM}{: (datatype) Description of structure. What it represents.}

\item{mY}{: (datatype) Description of structure. What it represents.}

\item{famX}{: (datatype) Description of structure. What it represents.}

\item{famM}{: (datatype) Description of structure. What it represents.}

\item{famY}{: (datatype) Description of structure. What it represents.}
}
\value{
: (vector) a length-2 list with the CDE with RR-interaction and
  RD-interaction.
}
\description{
Controlled Direct Effect (CDE) of the G-estimation model.
}
\examples{
'examples/cde_example.R'
}
\references{
Naimi, A. I., Schnitzer, M. E., Moodie, E. E. M., & Bodnar, L. M.
  (2016). Mediation Analysis for Health Disparities Research. American
  Journal of Epidemiology, 184(4), 315–324.
  https://doi.org/10.1093/aje/kwv329
}
\author{
Ashley Naimi \email{ashley.naimi@pitt.edu}

Cantwell Carson \email{carsonc@gmail.com}
}
