% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_cde}
\alias{compare_cde}
\title{Controlled Direct Effect (CDE) model comparison}
\usage{
compare_cde(x = NULL, m = NULL, c_xy = NULL, c_my = NULL, y = NULL,
  mX = NULL, mM = NULL, mY = NULL, famX = "gaussian",
  famM = "gaussian", famY = "gaussian", boot = TRUE, sims = 100,
  ci = 0.95)
}
\arguments{
\item{x}{: (datatype) Description of structure. What it represents.}

\item{m}{: (datatype) Description of structure. What it represents.}

\item{c_xy}{: (datatype) Description of structure. What it represents.}

\item{c_my}{: (datatype) Description of structure. What it represents.}

\item{y}{: (datatype) Description of structure. What it represents.}

\item{mX}{: (datatype) Description of structure. What it represents.}

\item{mM}{: (datatype) Description of structure. What it represents.}

\item{mY}{: (datatype) Description of structure. What it represents.}

\item{famX}{: (datatype) Description of structure. What it represents.}

\item{famM}{: (datatype) Description of structure. What it represents.}

\item{famY}{: (datatype) Description of structure. What it represents.}

\item{boot}{: (boolean) A boolean variable that indicates whether a
non-parametric bootstrap will be used to calculate confidence intervals}

\item{sims}{: (positive integer) If boot is TRUE, sims is the number of
simulations used to generate the confidence intervals.}

\item{ci}{: (float) A float between 0 and 1 representing the desired
confidence interval to be used for the non-parametric bootstrap.}
}
\value{
: (data frame) a data frame with all four mediation methods for
  compaisionm, with RR-interaction and RD-interaction. If boot is TRUE, it
  will also return the right and left bounds of the desired confidence
  interval, ci.
}
\description{
Controlled Direct Effect (CDE) model comparison
}
\examples{
'examples/cde_example.R'
}
\references{
Naimi, A. I., Schnitzer, M. E., Moodie, E. E. M., & Bodnar, L. M.
  (2016). Mediation Analysis for Health Disparities Research. American
  Journal of Epidemiology, 184(4), 315–324.
  https://doi.org/10.1093/aje/kwv329
}
\author{
Ashley Naimi \email{ashley.naimi@pitt.edu}

Cantwell Carson \email{carsonc@gmail.com}
}
