% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_boot}
\alias{get_boot}
\title{This is an an internal function called \code{get_boot()}}
\usage{
get_boot(cde_ci, ci)
}
\arguments{
\item{cde_ci}{: a dataframe containing the data necessary}

\item{ci}{: (float) A number between zero and 1 denoting the boot
interval desired}
}
\value{
a number corresponding to the normal boot half-interval for the
  resampled dataset
}
\description{
This returns a boot interval from a data set using manual bootstrap resampling.
}
\keyword{internal}
