% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_output}
\alias{get_output}
\title{This is an an internal function called \code{get_output()}}
\usage{
get_output(cde_ci, ci)
}
\arguments{
\item{ci}{: (float) A number between zero and 1 denoting the boot
interval desired}

\item{boot}{: (boolean) whether or not the boot interval should be calculated}

\item{df}{: a dataframe containing the data necessary}
}
\value{
a cbind list of risk ratio and risk difference along with their boot intervals
}
\description{
This formats a controlled direct effect estimate for risk difference and
risk ratio along with the boot intervals from resampling (if requested).
}
\keyword{internal}
