% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_propensity}
\alias{get_propensity}
\title{This is an an internal function called \code{get_propensity()}}
\usage{
get_propensity(data, mX, mM, famX, famM)
}
\arguments{
\item{data}{: (datatype) Description of structure. What it represents.}

\item{mX}{: (datatype) Description of structure. What it represents.}

\item{mM}{: (datatype) Description of structure. What it represents.}

\item{famX}{: (datatype) Description of structure. What it represents.}

\item{famM}{: (datatype) Description of structure. What it represents.}
}
\value{
a list of two propensities
}
\description{
This returns the propensities for exposure and mediator
}
\keyword{internal}
