% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_resample}
\alias{get_resample}
\title{This is an an internal function called \code{get_resample()}}
\usage{
get_resample(df)
}
\arguments{
\item{df}{: (dataframe) An arbitrary dataframe to be resampled.}
}
\value{
a list of two propensities
}
\description{
This returns the rows of a data frame resampled with repetition for
bootstrapping boot intervals.
}
\keyword{internal}
