% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_vars}
\alias{get_vars}
\title{This is an an internal function called \code{get_vars()}}
\usage{
get_vars(x, m, c_xy, c_my, y, mX, mM, mY)
}
\arguments{
\item{x}{: (datatype) Description of structure. What it represents.}

\item{m}{: (datatype) Description of structure. What it represents.}

\item{c_xy}{: (datatype) Description of structure. What it represents.}

\item{c_my}{: (datatype) Description of structure. What it represents.}

\item{y}{: (datatype) Description of structure. What it represents.}

\item{mX}{: (datatype) Description of structure. What it represents.}

\item{mM}{: (datatype) Description of structure. What it represents.}

\item{mY}{: (datatype) Description of structure. What it represents.}
}
\value{
a list composed of one data frame and three vectors
}
\description{
This returns a dataframe from vectors for exposure, mediator, c_xy, c_my, and
outcome as well as the individual vectors for exposure, mediator, and
outcome. It also checks to see if there are NULL values forany of these and
stops the program if any values are NULL.
}
\keyword{internal}
