% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ipw_f}
\alias{ipw_f}
\title{CounterfactuaL Disparity Measure (CDM) of the inverse probability-
weighted (IPW) marginal structural model.}
\usage{
ipw_f(outcome, mY = NULL, famY = NULL, mediator, mM, famM, exposure, mX,
  famX, data, interactionRR, interactionRD)
}
\arguments{
\item{outcome}{: (datatype) Description of structure. What it represents.}

\item{mY}{: (datatype) Description of structure. What it represents.}

\item{famY}{: (datatype) Description of structure. What it represents.}

\item{mediator}{: (datatype) Description of structure. What it represents.}

\item{mM}{: (datatype) Description of structure. What it represents.}

\item{famM}{: (datatype) Description of structure. What it represents.}

\item{exposure}{: (datatype) Description of structure. What it represents.}

\item{mX}{: (datatype) Description of structure. What it represents.}

\item{famX}{: (datatype) Description of structure. What it represents.}

\item{data}{: (datatype) Description of structure. What it represents.}

\item{interactionRR}{: (datatype) Description of structure. What it
represents.}

\item{interactionRD}{: (datatype) Description of structure. What it
represents.}
}
\value{
: (vector) a length-2 list with the CDM with RR-interaction and
  RD-interaction.
}
\description{
CounterfactuaL Disparity Measure (CDM) of the inverse probability-
weighted (IPW) marginal structural model.
}
\examples{
see cde_example.R
}
\references{
Naimi, A. I., Schnitzer, M. E., Moodie, E. E. M., & Bodnar, L. M.
  (2016). Mediation Analysis for Health Disparities Research. American
  Journal of Epidemiology, 184(4), 315–324.
  https://doi.org/10.1093/aje/kwv329
}
\author{
Ashley Naimi \email{ashley.naimi@pitt.edu}

Cantwell Carson \email{carsonc@gmail.com}
}
