% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimed_example.R
\docType{data}
\name{trimed_example}
\alias{trimed_example}
\title{This is random data for testing purposes to be included in my package}
\format{A data frame with 5000 rows and 5 variables:
\describe{
  \item{x}{exposure}
  \item{m}{mediator}
  \item{c_my}{Ashley, write something here}
  \item{c_xy}{Ashley, write something here}
  \item{y}{outcome}
  }}
\description{
This is random data for testing purposes to be included in my package
}
\author{
@author Ashley Naimi \email{ashley.naimi@pitt.edu}
}
\keyword{data}
